Feedback on PR incorporated: - comment removed - test added for coverage, tested flyTo now with coordinates - eslint applied for order of imports - added dependencies in useEffects
