+++ src/signals/IncidentMap/components/GPSLocation/index.ts
@@ -1,0 +1,1 @@
+export { GPSLocation } from './GPSLocation'
