--- src/signals/IncidentMap/components/ZoomToGPSLocation/index.ts
@@ -1,1 +1,0 @@
-export { ZoomToGPSLocation } from './ZoomToGPSLocation'
